+++ mrv2/docs/HISTORY.md
@@ -1,6 +1,9 @@
 v0.3.7
 ------
 - Added a half OCIO default to handle OpenEXR half images.
+- Added timeRange and inOutRange to timeline module.
+- Added setIn() and setOut() to timeline module to set the in and out time/frame/seconds.
+- Fixed timeRange conversion to string (__str__) and repr (__repr__).
 
 v0.3.6
 ------